% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{anomalies_from_inspect}
\alias{anomalies_from_inspect}
\title{Takes in an Inspect result and spits out an object that can be printed
by plot_capacc_external}
\usage{
anomalies_from_inspect(inspectres, x, tol = 1)
}
\description{
Takes in an Inspect result and spits out an object that can be printed
by plot_capacc_external
}
