% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{detect_anomalies_capacc}
\alias{detect_anomalies_capacc}
\title{Detect anomalies by CAPA-CC}
\usage{
detect_anomalies_capacc(x, est_band = 2, b = 1, b_point = 1, minsl = 2)
}
\arguments{
\item{x}{Data matrix.}
}
\value{
A capacc object.
}
\description{
Detect anomalies by CAPA-CC
}
