% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model.R
\name{linear_model}
\alias{linear_model}
\title{Train a linear model}
\usage{
linear_model(x, response, covariates)
}
\arguments{
\item{x}{Data matrix.}
}
\value{
An lm object.
}
\description{
Train a linear model
}
