% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_synthetic_data}
\alias{read_synthetic_data}
\title{Read some synthetic data.}
\usage{
read_synthetic_data(n = 1000, p_response = 5, p_covariates = 5)
}
\value{
A data.table with the synthetic data.
}
\description{
Read some synthetic data.
}
